Refactor file utils
