+++ config.go
@@ -1,6 +1,6 @@
 package main
 
-import "os/user"
+import "github.com/jan25/termracer/pkg/utils"
 
 const (
 	// TopLevelDir is name of directory that stores data for termracer
@@ -9,6 +9,9 @@ const (
 	// SamplesDir is a sub directory to store paragraph samples data
 	SamplesDir = "/samples/use"
 
+	// SamplesJSONPath is a file storing sample paragraphs in JSON format
+	SamplesJSONPath = SamplesDir + "/samples.json"
+
 	// HistoryFile is name of file to store race history data
 	HistoryFile = "/racehistory.csv"
 )
@@ -16,7 +19,7 @@ const (
 // GetSamplesUseDir returns path to samples/use
 // directory in local filesystem
 func GetSamplesUseDir() (string, error) {
-	home, err := getHomeDir()
+	home, err := utils.GetHomeDir()
 	if err != nil {
 		return "", err
 	}
@@ -25,7 +28,7 @@ func GetSamplesUseDir() (string, error) {
 
 // GetHistoryFilePath returns path to race history file
 func GetHistoryFilePath() (string, error) {
-	home, err := getHomeDir()
+	home, err := utils.GetHomeDir()
 	if err != nil {
 		return "", err
 	}
@@ -34,17 +37,9 @@ func GetHistoryFilePath() (string, error) {
 
 // GetTopLevelDir returns full path to top level dir
 func GetTopLevelDir() (string, error) {
-	home, err := getHomeDir()
+	home, err := utils.GetHomeDir()
 	if err != nil {
 		return "", err
 	}
 	return home + TopLevelDir, nil
-}
-
-func getHomeDir() (string, error) {
-	usr, err := user.Current()
-	if err != nil {
-		return "", err
-	}
-	return usr.HomeDir, nil
 }
